refactor(agent-core): split audit-inventory.ts into 4 focused modules
Closes #2646
